test(arel): move Array-name quoting tests into to-sql.trails.test.ts
